feat: working alias command replacement
